Remove catching Throwable exceptions
This fixes the TooGenericExceptionCaught linter issue. We ignored it, but it actually ended up causing issues. When a coroutine is cancelled it often throws a CancellationException. This was caught and converted to an ApiClientException causing coroutine cancellation to stop working. This change removes the places that catch Throwable, and they may throw exceptions now. We'll need to watch this and possibly add new catch branches later.