Optionally follow a redirect from the server in discovery if it's a subpath of the original.
Adds a parameter to RecommendedServerDiscovery.discover() that indicates if we should follow a redirect from the given servers.
If true, send a HEAD request to the given servers, and if it returns a redirect code then add the redirect location to the collection of servers to check, but only if the redirected location is on the same host as the original.
Prefer non-redirected addresses over those that are the result of a redirect.
Pass the original and possible redirected address back through the RecommendedServerInfo object for use.